requirements.txt
wheel
python-dotenv
# part 1
openai
groq
ollama
#geopy
langchain
langchain-ollama
langchain-groq
streamlit
llama-index